requirements.txt
streamlit==1.32.0
anthropic==0.25.0
chromadb==0.4.24
bcrypt==4.1.2
PyJWT==2.8.0
python-dotenv==1.0.1
pandas==2.2.0
